requirements.txt
# Django & REST Framework
Django
djangorestframework

# JWT Authentication
djangorestframework-simplejwt

# PostgreSQL driver
psycopg2-binary

# CORS handling (useful when Vue frontend calls backend)
django-cors-headers

# Environment variables support
python-decouple

# Soft delete functionality for Django models
django-soft-delete

#
drf-yasg

